angular wrapper - part 2
more #2310
* move demo code to angular new folder, updated links
