+++ angular/projects/demo/src/app/ngFor.ts
