+++ angular/projects/demo/src/app/ngFor_cmd.ts
