+++ angular/projects/demo/src/styles.css
@@ -1,1 +1,4 @@
-/* You can add global styles to this file, and also import other style files */
+/* re-use existing demo css file we already use for the plain demos - that include gridstack.css which is required */
+@import "../../../../demo/demo.css";
+/* required file for gridstack 2-11 column */
+@import "../../../../dist/gridstack-extra.css";
