--- demo/angular/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "angular-ng-for",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^15.1.1",
-    "@angular/common": "^15.1.1",
-    "@angular/compiler": "^15.1.1",
-    "@angular/core": "^15.1.1",
-    "@angular/forms": "^15.1.1",
-    "@angular/platform-browser": "^15.1.1",
-    "@angular/platform-browser-dynamic": "^15.1.1",
-    "@angular/router": "^15.1.1",
-    "gridstack": "^8.1.1",
-    "rxjs": "~7.8.1",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.13.0"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.2",
-    "@angular/cli": "^15.1.2",
-    "@angular/compiler-cli": "^15.1.1",
-    "@types/jasmine": "~4.3.1",
-    "@types/node": "^18.16.3",
-    "jasmine-core": "~4.6.0",
-    "karma": "~6.4.2",
-    "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "<5.0"
-  }
-}
