--- demo/angular/src/app/app.component.css
@@ -1,11 +1,0 @@
-.test-container {
-  margin-top: 30px;
-}
-button.active {
-  color: #fff;
-  background-color: #007bff;
-}
-
-.text-container {
-  display: flex;
-}
