--- demo/angular/tsconfig.lib.json
@@ -1,13 +1,0 @@
-/* config to publish the angular wrapper classes */
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": ".../../../../dist/ng",
-    "declaration": true,
-  },
-  "include": [
-    "./src/gridstack.component.ts",
-    "./src/gridstack-item.component.ts",
-    "./src/gridstack.module.ts",
-  ],
-}
