+++ demo/index.html
@@ -38,14 +38,14 @@
     <li><a href="esmodule.html">ES Module test</a></li>
   </ul>
   <h1>Angular wrapper</h1>
-  <p>We now ship an <a href="https://github.com/gridstack/gridstack.js/tree/master/demo/angular/" target="_blank">Angular component wrapper</a>
+  <p>We now ship an <a href="https://github.com/gridstack/gridstack.js/tree/master/angular/" target="_blank">Angular Component</a>
     to make it supper easy for that framework (Vue and React examples are above)</p>
-  <p>These are complete Angular projects with multiple options. use `yarn` and `yarn start` in <a href="https://github.com/gridstack/gridstack.js/tree/master/demo/angular" target="_blank">demo/angular</a> sub-project to run them</p>
+  <p>These are complete Angular projects with multiple options. use `yarn` and `yarn start` in <a href="https://github.com/gridstack/gridstack.js/tree/master/angular/projects/demo" target="_blank">angular demo</a> sub-project to run them</p>
   <ol>
-    <li><a href="angular/src/app/simple.ts">simple.ts</a></li>
-    <li><a href="angular/src/app/ngFor.ts">ngFor.ts</a></li>
-    <li><a href="angular/src/app/ngFor_cmd.ts">ngFor with command</a> (not recommended)</li>
-    <li><a href="angular/src/gridstack.component.ts">gridstack.component.ts</a> and <a href="angular/src/gridstack-item.component.ts">gridstack-item.component.ts</a> (BEST)</li>
+    <li><a href="../angular/projects/demo/src/app/simple.ts">simple.ts</a></li>
+    <li><a href="../angular/projects/demo/src/app/ngFor.ts">ngFor.ts</a></li>
+    <li><a href="../angular/projects/demo/src/app/ngFor_cmd.ts">ngFor with command</a> (not recommended)</li>
+    <li><a href="../angular/projects/lib/src/lib/gridstack.component.ts">gridstack.component.ts</a> and <a href="../angular/projects/lib/src/lib/gridstack-item.component.ts">gridstack-item.component.ts</a> (BEST)</li>
   </ol>
   <h1>Old v5.1.1 Jquery Demos</h1>
   Note: those are no longer supported, and use an old version of the lib to compare functionality.
